requirements.txt
# freqtrade
ccxt
ta
# pandas-ta
openai
groq
numpy
pandas
scikit-learn
# matplotlib
# seaborn
python-dotenv
requests
pytest
pytest-asyncio
optuna>=3.0.0
# TA-Lib - Installation attempted, but failed due to C library / Python wrapper compatibility issues.
# Requires manual resolution or a different C library source / Python wrapper version.
# TA-Lib
